pyproject.toml
[project]
name = "test-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.0.1",
]
